requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
coveopush==0.2
cryptography==3.3.1
cssselect==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
jmespath==0.10.0
jsonpickle==1.5.1
lxml==4.6.2
parsel==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
soupsieve==2.1
urllib3==1.26.3
w3lib==1.22.0
zope.interface==5.2.0
